# Faster `take` with null values (2-3x)