--- a/src/compute/take/boolean.rs
+++ b/src/compute/take/boolean.rs
@@ -1,20 +1,3 @@
-// Licensed to the Apache Software Foundation (ASF) under one
-// or more contributor license agreements.  See the NOTICE file
-// distributed with this work for additional information
-// regarding copyright ownership.  The ASF licenses this file
-// to you under the Apache License, Version 2.0 (the
-// "License"); you may not use this file except in compliance
-// with the License.  You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-
 use crate::{
     array::{Array, BooleanArray, PrimitiveArray},
     bitmap::{Bitmap, MutableBitmap},
@@ -35,21 +18,16 @@ fn take_values_validity<I: Index>(
     values: &BooleanArray,
     indices: &[I],
 ) -> (Bitmap, Option<Bitmap>) {
-    let mut validity = MutableBitmap::with_capacity(indices.len());
-
     let validity_values = values.validity().unwrap();
+    let validity = indices
+        .iter()
+        .map(|index| validity_values.get_bit(index.to_usize()));
+    let validity = Bitmap::from_trusted_len_iter(validity);
 
     let values_values = values.values();
-
-    let values = indices.iter().map(|index| {
-        let index = index.to_usize();
-        if validity_values.get_bit(index) {
-            validity.push(true);
-        } else {
-            validity.push(false);
-        }
-        values_values.get_bit(index)
-    });
+    let values = indices
+        .iter()
+        .map(|index| values_values.get_bit(index.to_usize()));
     let buffer = Bitmap::from_trusted_len_iter(values);
 
     (buffer, validity.into())