@ritchie46 2021-11-25
Intersting. I also had a similar result with splitting the bound check in two loops.

Maybe that makes sense here as well. Als it is now checked in the validity and in the values.